+++ src/mnemonic.hpp
@@ -13,11 +13,12 @@
 #include "mnemonic_djnz.hpp"
 #include "mnemonic_ex.hpp"
 #include "mnemonic_im.hpp"
+#include "mnemonic_in.hpp"
 #include "mnemonic_inc.hpp"
-#include "mnemonic_io.hpp"
 #include "mnemonic_jp.hpp"
 #include "mnemonic_jr.hpp"
 #include "mnemonic_load.hpp"
+#include "mnemonic_out.hpp"
 #include "mnemonic_rst.hpp"
 #include "mnemonic_shift.hpp"
 #include "mnemonic_stack.hpp"
